refactor: move dtos to separate file
